test(worker): run NaN detection tests on CPU
